Move type-related classes under the `type` package
