+++ client/src/main/java/org/spine3/base/Events.java
@@ -24,7 +24,7 @@ import com.google.protobuf.Any;
 import com.google.protobuf.Message;
 import com.google.protobuf.Timestamp;
 import org.spine3.protobuf.Timestamps2;
-import org.spine3.protobuf.TypeName;
+import org.spine3.type.TypeName;
 import org.spine3.users.UserId;
 
 import java.util.Collections;
