+++ client/src/main/java/org/spine3/protobuf/AnyPacker.java
@@ -23,7 +23,8 @@ package org.spine3.protobuf;
 import com.google.protobuf.Any;
 import com.google.protobuf.InvalidProtocolBufferException;
 import com.google.protobuf.Message;
-import org.spine3.protobuf.error.UnexpectedTypeException;
+import org.spine3.type.TypeUrl;
+import org.spine3.type.error.UnexpectedTypeException;
 
 import java.util.Iterator;
 
@@ -87,7 +88,7 @@ public class AnyPacker {
      * Unwraps {@code Any} value into an instance of the passed class.
      *
      * <p>If there is no Java class for the type,
-     * {@link org.spine3.protobuf.error.UnexpectedTypeException UnexpectedTypeException}
+     * {@link UnexpectedTypeException UnexpectedTypeException}
      * will be thrown.
      *
      * @param any   instance of {@link Any} that should be unwrapped
