+++ client/src/main/java/org/spine3/type/TypeName.java
@@ -18,7 +18,7 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package org.spine3.protobuf;
+package org.spine3.type;
 
 import com.google.protobuf.Descriptors.Descriptor;
 import com.google.protobuf.Message;
