+++ client/src/main/java/org/spine3/type/error/UnknownTypeException.java
@@ -17,7 +17,7 @@
  * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-package org.spine3.protobuf.error;
+package org.spine3.type.error;
 
 /**
  * Exception that is thrown when unsupported message is obtained
