+++ client/src/test/java/org/spine3/protobuf/MessagesShould.java
@@ -38,6 +38,8 @@ import org.spine3.test.Tests;
 import org.spine3.test.messages.MessageWithStringValue;
 import org.spine3.test.messages.TestEnum;
 import org.spine3.test.messages.TestEnumValue;
+import org.spine3.type.KnownTypes;
+import org.spine3.type.TypeUrl;
 
 import java.util.List;
