+++ client/src/test/java/org/spine3/type/TypeUrlShould.java
@@ -18,7 +18,7 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package org.spine3.protobuf;
+package org.spine3.type;
 
 import com.google.protobuf.Any;
 import com.google.protobuf.Descriptors;
@@ -38,10 +38,10 @@ import org.spine3.validate.internal.IfMissingOption;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertTrue;
 import static org.spine3.base.Identifiers.newUuid;
-import static org.spine3.protobuf.TypeUrl.GOOGLE_TYPE_URL_PREFIX;
-import static org.spine3.protobuf.TypeUrl.SPINE_TYPE_URL_PREFIX;
-import static org.spine3.protobuf.TypeUrl.composeTypeUrl;
 import static org.spine3.protobuf.Values.newStringValue;
+import static org.spine3.type.TypeUrl.GOOGLE_TYPE_URL_PREFIX;
+import static org.spine3.type.TypeUrl.SPINE_TYPE_URL_PREFIX;
+import static org.spine3.type.TypeUrl.composeTypeUrl;
 
 public class TypeUrlShould {
