+++ server/src/main/java/org/spine3/server/command/CommandStorage.java
@@ -32,8 +32,8 @@ import org.spine3.base.CommandStatus;
 import org.spine3.base.Commands;
 import org.spine3.base.Error;
 import org.spine3.base.Failure;
-import org.spine3.protobuf.TypeUrl;
 import org.spine3.server.storage.AbstractStorage;
+import org.spine3.type.TypeUrl;
 
 import javax.annotation.Nullable;
 import java.util.Iterator;
