+++ server/src/main/java/org/spine3/server/entity/FieldMasks.java
@@ -26,8 +26,8 @@ import com.google.protobuf.Message;
 import com.google.protobuf.ProtocolStringList;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import org.spine3.protobuf.KnownTypes;
-import org.spine3.protobuf.TypeUrl;
+import org.spine3.type.KnownTypes;
+import org.spine3.type.TypeUrl;
 
 import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
