+++ server/src/main/java/org/spine3/server/event/enrich/EventEnricher.java
@@ -38,9 +38,9 @@ import org.spine3.base.Enrichment.Container;
 import org.spine3.base.Event;
 import org.spine3.base.EventContext;
 import org.spine3.protobuf.AnyPacker;
-import org.spine3.protobuf.TypeName;
-import org.spine3.protobuf.TypeUrl;
 import org.spine3.type.EventClass;
+import org.spine3.type.TypeName;
+import org.spine3.type.TypeUrl;
 
 import javax.annotation.Nullable;
 import java.util.Collection;
