+++ server/src/main/java/org/spine3/server/event/error/EventException.java
@@ -27,7 +27,7 @@ import com.google.protobuf.Message;
 import com.google.protobuf.Value;
 import org.spine3.base.Error;
 import org.spine3.protobuf.AnyPacker;
-import org.spine3.protobuf.TypeName;
+import org.spine3.type.TypeName;
 
 import java.util.Map;
