+++ server/src/main/java/org/spine3/server/event/error/UnsupportedEventException.java
@@ -22,8 +22,8 @@ package org.spine3.server.event.error;
 import com.google.protobuf.Message;
 import org.spine3.base.Error;
 import org.spine3.base.EventValidationError;
-import org.spine3.protobuf.TypeName;
 import org.spine3.type.EventClass;
+import org.spine3.type.TypeName;
 
 /**
  * Exception that is thrown when unsupported event is obtained
