+++ server/src/main/java/org/spine3/server/storage/RecordStorage.java
@@ -25,11 +25,11 @@ import com.google.protobuf.Any;
 import com.google.protobuf.FieldMask;
 import com.google.protobuf.Message;
 import org.spine3.protobuf.AnyPacker;
-import org.spine3.protobuf.TypeUrl;
 import org.spine3.server.entity.EntityRecord;
 import org.spine3.server.entity.FieldMasks;
 import org.spine3.server.entity.LifecycleFlags;
 import org.spine3.server.stand.AggregateStateId;
+import org.spine3.type.TypeUrl;
 
 import java.util.Map;
