+++ server/src/test/java/org/spine3/server/Given.java
@@ -27,7 +27,6 @@ import org.spine3.base.Commands;
 import org.spine3.base.Identifiers;
 import org.spine3.client.Target;
 import org.spine3.people.PersonName;
-import org.spine3.protobuf.TypeName;
 import org.spine3.server.aggregate.Aggregate;
 import org.spine3.server.aggregate.AggregateRepository;
 import org.spine3.server.aggregate.Apply;
@@ -47,6 +46,7 @@ import org.spine3.test.commandservice.customer.command.CreateCustomer;
 import org.spine3.test.commandservice.customer.event.CustomerCreated;
 import org.spine3.time.LocalDate;
 import org.spine3.time.LocalDates;
+import org.spine3.type.TypeName;
 import org.spine3.users.UserId;
 
 import java.util.List;
