+++ server/src/test/java/org/spine3/testdata/Sample.java
@@ -36,9 +36,9 @@ import org.spine3.base.Event;
 import org.spine3.base.EventContext;
 import org.spine3.base.Events;
 import org.spine3.protobuf.AnyPacker;
-import org.spine3.protobuf.KnownTypes;
-import org.spine3.protobuf.TypeUrl;
 import org.spine3.type.ClassName;
+import org.spine3.type.KnownTypes;
+import org.spine3.type.TypeUrl;
 
 import java.lang.reflect.InvocationTargetException;
 import java.lang.reflect.Method;
